Rename the eyb app to investment_lead
